gke-tpu-examples/single-host-inference/jax/jax-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt requirements.in
#
--find-links https://storage.googleapis.com/jax-releases/libtpu_releases.html

absl-py==1.4.0
    # via
    #   chex
    #   optax
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
chex==0.1.7
    # via optax
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
diffusers==0.15.0
    # via -r requirements.in
dm-tree==0.1.8
    # via chex
filelock==3.10.7
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
flatbuffers==23.3.3
    # via tensorflow
flax==0.6.2
    # via -r requirements.in
fonttools==4.39.3
    # via matplotlib
gast==0.4.0
    # via tensorflow
google-auth==2.17.2
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.53.0
    # via
    #   tensorboard
    #   tensorflow
    #   tensorflow-serving-api
h5py==3.8.0
    # via tensorflow
huggingface-hub==0.13.3
    # via
    #   diffusers
    #   transformers
idna==3.4
    # via requests
importlib-metadata==6.3.0
    # via diffusers
jax[tpu]==0.4.6
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   optax
jaxlib==0.4.6
    # via
    #   chex
    #   jax
    #   optax
keras==2.11.0
    # via tensorflow
kiwisolver==1.4.4
    # via matplotlib
libclang==16.0.0
    # via tensorflow
libtpu-nightly==0.1.dev20230309
    # via jax
markdown==3.4.3
    # via tensorboard
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via werkzeug
matplotlib==3.7.1
    # via flax
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.5
    # via flax
numpy==1.24.2
    # via
    #   chex
    #   contourpy
    #   diffusers
    #   flax
    #   h5py
    #   jax
    #   jaxlib
    #   matplotlib
    #   opt-einsum
    #   optax
    #   scipy
    #   tensorboard
    #   tensorflow
    #   tensorstore
    #   transformers
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via
    #   jax
    #   tensorflow
optax==0.1.4
    # via flax
packaging==23.0
    # via
    #   huggingface-hub
    #   matplotlib
    #   tensorflow
    #   transformers
pbr==5.11.1
    # via testresources
pillow==9.5.0
    # via
    #   diffusers
    #   matplotlib
protobuf==3.19.6
    # via
    #   tensorboard
    #   tensorflow
    #   tensorflow-serving-api
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.14.0
    # via rich
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via
    #   flax
    #   huggingface-hub
    #   transformers
regex==2023.3.23
    # via
    #   diffusers
    #   transformers
requests==2.28.2
    # via
    #   diffusers
    #   huggingface-hub
    #   jax
    #   requests-oauthlib
    #   tensorboard
    #   transformers
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==13.3.3
    # via flax
rsa==4.9
    # via google-auth
scipy==1.10.1
    # via
    #   jax
    #   jaxlib
six==1.16.0
    # via
    #   astunparse
    #   google-auth
    #   google-pasta
    #   python-dateutil
    #   tensorflow
tensorboard==2.11.2
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.11.0
    # via
    #   -r requirements.in
    #   tensorflow-serving-api
tensorflow-estimator==2.11.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.32.0
    # via tensorflow
tensorflow-serving-api==2.11.0
    # via -r requirements.in
tensorstore==0.1.35
    # via flax
termcolor==2.2.0
    # via tensorflow
testresources==2.0.1
    # via -r requirements.in
tokenizers==0.13.3
    # via transformers
toolz==0.12.0
    # via chex
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   transformers
transformers==4.27.3
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   chex
    #   flax
    #   huggingface-hub
    #   optax
    #   tensorflow
urllib3==1.26.15
    # via requests
werkzeug==2.2.3
    # via tensorboard
wheel==0.40.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.15.0
    # via tensorflow
zipp==3.1.0
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==67.6.1
    # via
    #   tensorboard
    #   tensorflow
